+++ engine/tests/new-tests/filter_tests.cu
@@ -4,12 +4,11 @@
 #include <cudf/sorting.hpp>
 #include <cudf/scalar/scalar_factories.hpp>
 
-#include "from_cudf/cpp_tests/utilities/base_fixture.hpp"
-#include "from_cudf/cpp_tests/utilities/column_utilities.hpp"
-#include "from_cudf/cpp_tests/utilities/column_wrapper.hpp"
-// #include "from_cudf/cpp_tests/utilities/legacy/cudf_test_utils.cuh"
-#include "from_cudf/cpp_tests/utilities/table_utilities.hpp"
-#include "from_cudf/cpp_tests/utilities/type_lists.hpp"
+#include "tests/utilities/base_fixture.hpp"
+#include "tests/utilities/column_utilities.hpp"
+#include "tests/utilities/column_wrapper.hpp"
+#include "tests/utilities/table_utilities.hpp"
+#include "tests/utilities/type_lists.hpp"
 #include "execution_graph/logic_controllers/LogicalFilter.h"
 #include "utilities/DebuggingUtils.h"
 #include "tests/BlazingUnitTest.h"
@@ -98,7 +97,7 @@ TYPED_TEST(LogicalFilterTest, filter_table)
                                                   nullptr);
 
   cudf::size_type outputRows = (inputRows / 2);
-  if (cudf::type_to_id<T>() == cudf::BOOL8) {
+  if (cudf::type_to_id<T>() == cudf::type_id::BOOL8) {
     outputRows = inputRows;
   }