+++ engine/tests/new-tests/interops_tests.cu
@@ -4,12 +4,12 @@
 #include <cudf/sorting.hpp>
 #include <cudf/scalar/scalar_factories.hpp>
 
-#include "from_cudf/cpp_tests/utilities/base_fixture.hpp"
-#include "from_cudf/cpp_tests/utilities/column_utilities.hpp"
-#include "from_cudf/cpp_tests/utilities/column_wrapper.hpp"
-#include "from_cudf/cpp_tests/utilities/table_utilities.hpp"
-#include "from_cudf/cpp_tests/utilities/type_lists.hpp"
-#include "from_cudf/cpp_tests/utilities/type_list_utilities.hpp"
+#include "tests/utilities/base_fixture.hpp"
+#include "tests/utilities/column_utilities.hpp"
+#include "tests/utilities/column_wrapper.hpp"
+#include "tests/utilities/table_utilities.hpp"
+#include "tests/utilities/type_lists.hpp"
+#include "tests/utilities/type_list_utilities.hpp"
 #include "Interpreter/interpreter_cpp.h"
 #include "tests/BlazingUnitTest.h"