+++ CHANGELOG.md
@@ -5,7 +5,9 @@
 
 ## Improvements
 - #777 Update Calcite to the most recent version 1.23
+- #822 remove "from_cudf" code and cudf test utilities from engine code
 - #824 Added a test on Calcite to compare the logical plans when the ruleset is updated
+
 
 
 
@@ -20,6 +22,7 @@
 - #799 Fix uri values computacion in runQueryCaller
 - #792 Remove orc temp files when cached on Disk
 - #814 Fix when checking only Limit and Scan Kernels
+- #816 Loading one file at a time (LimitKernel and ScanKernel) 
 
 # BlazingSQL 0.14.0 (June 9, 2020)